태그 추가 완성본.N:M 표현

+++ src/main/java/com/pozafly/tripllo/board/dao/BoardDao.java
@@ -15,7 +15,7 @@ public interface BoardDao {
     public List<Board> readRecentBoards(List<String> boardIdList);
     public List<Board> readInvitedBoards(String userId);
     public BoardResultMap readBoardDetail(Long boardId);
-    public int createBoard(Map<String, Object> boardInfo);
+    public Long createBoard(Map<String, Object> boardInfo);
     public void updateBoard(Map<String, Object> boardInfo);
     public void deleteBoard(Long boardId);
 }
